feat(chat): refactor chat message handling and introduce chat-message-state module
